Removed context, went with i18n
