+++ web/locale/en/messages.js
@@ -1,1 +1,1 @@
-module.exports={l:{"p":function(n,ord){var s=String(n).split("."),v0=!s[1],t0=Number(s[0])==n,n10=t0&&s[0].slice(-1),n100=t0&&s[0].slice(-2);if(ord)return n10==1&&n100!=11?"one":n10==2&&n100!=12?"two":n10==3&&n100!=13?"few":"other";return n==1&&v0?"one":"other"}},m:{"You can\u2019t select more than 3 days. To change your selections, remove some days first.":"You can\u2019t select more than 3 days. To change your selections, remove some days first.","Your 3 selected days:":"Your 3 selected days:","Remove day":"Remove day","Cancel request":"Cancel request","Please check these sections for errors:":"Please check these sections for errors:","Contact":"Contact","https://www.canada.ca/en/transparency/privacy.html":"https://www.canada.ca/en/transparency/privacy.html","Privacy":"Privacy","https://digital.canada.ca/legal/terms/":"https://digital.canada.ca/legal/terms/","Terms":"Terms","Request a new citizenship appointment":"Request a new citizenship appointment","Language Selection":"Language Selection","This is a new service, help us improve by":"This is a new service, help us improve by","https://docs.google.com/forms/d/e/1FAIpQLSdEF3D7QCZ1ecPVKdqz_-dQAvlVdwdCQtHHLzg_v2q5q7XBlg/viewform":"https://docs.google.com/forms/d/e/1FAIpQLSdEF3D7QCZ1ecPVKdqz_-dQAvlVdwdCQtHHLzg_v2q5q7XBlg/viewform","sending your feedback":"sending your feedback","BETA":"BETA","ALPHA":"ALPHA","Important message":"Important message","No days selected":"No days selected","Change":"Change","Full name":"Full name","Email":"Email","Paper file number":"Paper file number","Reason":"Reason","Explanation":"Explanation","Availability":"Availability","Travel":"Travel","Calendar":"Calendar","Go back":"Go back","Citizenship appointments are scheduled on Wednesdays and Thursdays.":"Citizenship appointments are scheduled on Wednesdays and Thursdays.","Select":"Select","3 days":"3 days","you\u2019re available between":"you\u2019re available between","and":"and","Make sure you stay available on all of the days you select.":"Make sure you stay available on all of the days you select.","Please select 2 more days to continue.":"Please select 2 more days to continue.","Please select 1 more day to continue.":"Please select 1 more day to continue.","Review request":"Review request","You must select 3 days.":"You must select 3 days.","Start over":"Start over","Your request has been cancelled.":"Your request has been cancelled.","Nothing about your application or appointment has been changed.":"Nothing about your application or appointment has been changed.","What happens next?":"What happens next?","Please attend your appointment at the date and time specified in the email we sent you.":"Please attend your appointment at the date and time specified in the email we sent you.","If you missed your appointment, you still need to contact IRCC within 30 days. If you don\u2019t, your application could be delayed or abandoned.":"If you missed your appointment, you still need to contact IRCC within 30 days. If you don\u2019t, your application could be delayed or abandoned.","If you have any questions, please contact:":"If you have any questions, please contact:","Sorry, something went wrong. We received your request, but you might not get a confirmation email. Please make note of your request information":"Sorry, something went wrong. We received your request, but you might not get a confirmation email. Please make note of your request information","Availability:":"Availability:","Thank you! Your request has been received.":"Thank you! Your request has been received.","We\u2019ve sent you a confirmation email.":"We\u2019ve sent you a confirmation email.","By":"By","your local":"your local","office will send you a new appointment, or email you to ask for more information.":"office will send you a new appointment, or email you to ask for more information.","Sorry, something went wrong.":"Sorry, something went wrong.","Your request":"Your request","was not completed":"was not completed","Your appointment or application wasn\u2019t changed in any way":"Your appointment or application wasn\u2019t changed in any way","Contact <0>IRCC</0> directly to reschedule your appointment:":"Contact <0>IRCC</0> directly to reschedule your appointment:","Our team has been notified of this error, but you can also <0>tell us what went wrong</0>.":"Our team has been notified of this error, but you can also <0>tell us what went wrong</0>.","Tell us what went wrong":"Tell us what went wrong","Home":"Home","Page not found":"Page not found","Sorry, the page you are looking for doesn\u2019t exist.":"Sorry, the page you are looking for doesn\u2019t exist.","Return to the home page to reschedule your appointment.":"Return to the home page to reschedule your appointment.","Tell IRCC you can\u2019t attend your citizenship appointment, and request a new one.":"Tell IRCC you can\u2019t attend your citizenship appointment, and request a new one.","You will need:":"You will need:","Your paper file number":"Your paper file number","To describe your reason for rescheduling":"To describe your reason for rescheduling","Then you\u2019ll select":"Then you\u2019ll select","between":"between","to":"to","when you\u2019re available for an appointment in the future":"when you\u2019re available for an appointment in the future","Requesting a new appointment will cancel your current one. Do not attend your old appointment after you complete this request. It can take up to 9 weeks for us to reschedule you.":"Requesting a new appointment will cancel your current one. Do not attend your old appointment after you complete this request. It can take up to 9 weeks for us to reschedule you.","Start now":"Start now","Email address":"Email address","Why are you rescheduling?":"Why are you rescheduling?","Describe why you can\u2019t attend your appointment":"Describe why you can\u2019t attend your appointment","Some information is missing.":"Some information is missing.","This is the full name you used on your citizenship application.":"This is the full name you used on your citizenship application.","This is where we\u2019ll send a confirmation email when you\u2019re done.":"This is where we\u2019ll send a confirmation email when you\u2019re done.","This number is at the top of the email attachment we sent you.":"This number is at the top of the email attachment we sent you.","Travel (business or vacation)":"Travel (business or vacation)","Medical":"Medical","Work or school":"Work or school","Family":"Family","Other":"Other","Provide enough detail so that staff can understand your situation.":"Provide enough detail so that staff can understand your situation.","Continue":"Continue","Send request":"Send request","Work or School":"Work or School","Review your request:":"Review your request:","Sending this request will cancel your current appointment.<0> Do not attend your old appointment</0> after you send this request.":"Sending this request will cancel your current appointment.<0> Do not attend your old appointment</0> after you send this request.","January":"January","February":"February","March":"March","April":"April","May":"May","June":"June","July":"July","August":"August","September":"September","October":"October","November":"November","December":"December","Sunday":"Sunday","Monday":"Monday","Tuesday":"Tuesday","Wednesday":"Wednesday","Thursday":"Thursday","Friday":"Friday","Saturday":"Saturday","Su":"Su","Mo":"Mo","Tu":"Tu","We":"We","Thu":"Thu","Fri":"Fri","Sat":"Sat","You need to tell us your name so we know who is requesting a new appointment.":"You need to tell us your name so we know who is requesting a new appointment.","Needs to be shorter than 20 words. Please use the name recorded on your application.":"Needs to be shorter than 20 words. Please use the name recorded on your application.","We need your email address so we can send you a confirmation message.":"We need your email address so we can send you a confirmation message.","Please make sure you provide a valid email address. For example, \u2018yourname@example.com\u2019.":"Please make sure you provide a valid email address. For example, \u2018yourname@example.com\u2019.","We need your paper file number so we can confirm your identity.":"We need your paper file number so we can confirm your identity.","Needs a number with a different format. Please make sure this is your correct Paper file number.":"Needs a number with a different format. Please make sure this is your correct Paper file number.","Please tell us why you need to reschedule your appointment. If none of the options fit your situation, choose \u2018Other\u2019.":"Please tell us why you need to reschedule your appointment. If none of the options fit your situation, choose \u2018Other\u2019.","Please tell us a bit more about why you need to reschedule your appointment.":"Please tell us a bit more about why you need to reschedule your appointment.","Sorry, there\u2018s a limit of 150 words for this explanation. Please shorten your explanation.":"Sorry, there\u2018s a limit of 150 words for this explanation. Please shorten your explanation.","Exactly three dates must be passed":"Exactly three dates must be passed","Reason needs to be on the list provided. Please pick one.":"Reason needs to be on the list provided. Please pick one."}};
+module.exports={l:{"p":function(n,ord){var s=String(n).split("."),v0=!s[1],t0=Number(s[0])==n,n10=t0&&s[0].slice(-1),n100=t0&&s[0].slice(-2);if(ord)return n10==1&&n100!=11?"one":n10==2&&n100!=12?"two":n10==3&&n100!=13?"few":"other";return n==1&&v0?"one":"other"}},m:{"Immigration, Refugees and Citizenship Canada":"Immigration, Refugees and Citizenship Canada","Our team has been notified of this error, but you can also":"Our team has been notified of this error, but you can also","let us know what happened":"let us know what happened","send an email to let us know what happened":"send an email to let us know what happened","You can\u2019t select more than 3 days. To change your selections, remove some days first.":"You can\u2019t select more than 3 days. To change your selections, remove some days first.","Your 3 selected days:":"Your 3 selected days:","Remove day":"Remove day","Cancel request":"Cancel request","Please check these sections for errors:":"Please check these sections for errors:","Contact":"Contact","https://www.canada.ca/en/transparency/privacy.html":"https://www.canada.ca/en/transparency/privacy.html","Privacy":"Privacy","https://digital.canada.ca/legal/terms/":"https://digital.canada.ca/legal/terms/","Terms":"Terms","Request a new citizenship appointment":"Request a new citizenship appointment","Language Selection":"Language Selection","This is a new service, help us improve by":"This is a new service, help us improve by","https://docs.google.com/forms/d/e/1FAIpQLSdEF3D7QCZ1ecPVKdqz_-dQAvlVdwdCQtHHLzg_v2q5q7XBlg/viewform":"https://docs.google.com/forms/d/e/1FAIpQLSdEF3D7QCZ1ecPVKdqz_-dQAvlVdwdCQtHHLzg_v2q5q7XBlg/viewform","sending your feedback":"sending your feedback","BETA":"BETA","ALPHA":"ALPHA","Important message":"Important message","No days selected":"No days selected","Send request":"Send request","Change":"Change","Full name":"Full name","Email":"Email","Paper file number":"Paper file number","Reason":"Reason","Explanation":"Explanation","Availability":"Availability","Travel":"Travel","Go back":"Go back","Citizenship appointments are scheduled on Wednesdays and Thursdays.":"Citizenship appointments are scheduled on Wednesdays and Thursdays.","Select":"Select","3 days":"3 days","you\u2019re available between":"you\u2019re available between","and":"and","Make sure you stay available on all of the days you select.":"Make sure you stay available on all of the days you select.","Please select 2 more days to continue.":"Please select 2 more days to continue.","Please select 1 more day to continue.":"Please select 1 more day to continue.","Review request":"Review request","You must select 3 days.":"You must select 3 days.","Start over":"Start over","Your request has been cancelled.":"Your request has been cancelled.","Nothing about your application or appointment has been changed.":"Nothing about your application or appointment has been changed.","What happens next?":"What happens next?","Please attend your appointment at the date and time specified in the email we sent you.":"Please attend your appointment at the date and time specified in the email we sent you.","If you missed your appointment, you still need to contact IRCC within 30 days. If you don\u2019t, your application could be delayed or abandoned.":"If you missed your appointment, you still need to contact IRCC within 30 days. If you don\u2019t, your application could be delayed or abandoned.","If you have any questions, please contact:":"If you have any questions, please contact:","Sorry, something went wrong. We received your request, but you might not get a confirmation email. Please make note of your request information":"Sorry, something went wrong. We received your request, but you might not get a confirmation email. Please make note of your request information","Availability:":"Availability:","Thank you! Your request has been received.":"Thank you! Your request has been received.","We\u2019ve sent you a confirmation email.":"We\u2019ve sent you a confirmation email.","By":"By","your local":"your local","office will send you a new appointment, or email you to ask for more information.":"office will send you a new appointment, or email you to ask for more information.","Sorry, something went wrong.":"Sorry, something went wrong.","Your request":"Your request","was not completed":"was not completed","Your appointment or application wasn\u2019t changed in any way":"Your appointment or application wasn\u2019t changed in any way","Contact <0>IRCC</0> directly to reschedule your appointment:":"Contact <0>IRCC</0> directly to reschedule your appointment:","Home":"Home","Page not found":"Page not found","Sorry, the page you are looking for doesn\u2019t exist.":"Sorry, the page you are looking for doesn\u2019t exist.","Return to the home page to reschedule your appointment.":"Return to the home page to reschedule your appointment.","Tell IRCC you can\u2019t attend your citizenship appointment, and request a new one.":"Tell IRCC you can\u2019t attend your citizenship appointment, and request a new one.","You will need:":"You will need:","Your paper file number":"Your paper file number","To describe your reason for rescheduling":"To describe your reason for rescheduling","Then you\u2019ll select":"Then you\u2019ll select","Requesting a new appointment will cancel your current one. Do not attend your old appointment after you complete this request. It can take up to 9 weeks for us to reschedule you.":"Requesting a new appointment will cancel your current one. Do not attend your old appointment after you complete this request. It can take up to 9 weeks for us to reschedule you.","Start now":"Start now","Email address":"Email address","Why are you rescheduling?":"Why are you rescheduling?","Describe why you can\u2019t attend your appointment":"Describe why you can\u2019t attend your appointment","Some information is missing.":"Some information is missing.","This is the full name you used on your citizenship application.":"This is the full name you used on your citizenship application.","This is where we\u2019ll send a confirmation email when you\u2019re done.":"This is where we\u2019ll send a confirmation email when you\u2019re done.","This number is at the top of the email attachment we sent you.":"This number is at the top of the email attachment we sent you.","Travel (business or vacation)":"Travel (business or vacation)","Medical":"Medical","Work or school":"Work or school","Family":"Family","Other":"Other","Provide enough detail so that staff can understand your situation.":"Provide enough detail so that staff can understand your situation.","Continue":"Continue","Work or School":"Work or School","Review your request:":"Review your request:","Sending this request will cancel your current appointment.<0> Do not attend your old appointment</0> after you send this request.":"Sending this request will cancel your current appointment.<0> Do not attend your old appointment</0> after you send this request.","January":"January","February":"February","March":"March","April":"April","May":"May","June":"June","July":"July","August":"August","September":"September","October":"October","November":"November","December":"December","Sunday":"Sunday","Monday":"Monday","Tuesday":"Tuesday","Wednesday":"Wednesday","Thursday":"Thursday","Friday":"Friday","Saturday":"Saturday","Su":"Su","Mo":"Mo","Tu":"Tu","We":"We","Thu":"Thu","Fri":"Fri","Sat":"Sat","You need to tell us your name so we know who is requesting a new appointment.":"You need to tell us your name so we know who is requesting a new appointment.","Needs to be shorter than 20 words. Please use the name recorded on your application.":"Needs to be shorter than 20 words. Please use the name recorded on your application.","We need your email address so we can send you a confirmation message.":"We need your email address so we can send you a confirmation message.","Please make sure you provide a valid email address. For example, \u2018yourname@example.com\u2019.":"Please make sure you provide a valid email address. For example, \u2018yourname@example.com\u2019.","We need your paper file number so we can confirm your identity.":"We need your paper file number so we can confirm your identity.","Needs a number with a different format. Please make sure this is your correct Paper file number.":"Needs a number with a different format. Please make sure this is your correct Paper file number.","Please tell us why you need to reschedule your appointment. If none of the options fit your situation, choose \u2018Other\u2019.":"Please tell us why you need to reschedule your appointment. If none of the options fit your situation, choose \u2018Other\u2019.","Please tell us a bit more about why you need to reschedule your appointment.":"Please tell us a bit more about why you need to reschedule your appointment.","Sorry, there\u2018s a limit of 150 words for this explanation. Please shorten your explanation.":"Sorry, there\u2018s a limit of 150 words for this explanation. Please shorten your explanation.","Exactly three dates must be passed":"Exactly three dates must be passed","Reason needs to be on the list provided. Please pick one.":"Reason needs to be on the list provided. Please pick one.","Calendar":"Calendar","Our team has been notified of this error, but you can also <0>tell us what went wrong</0>.":"Our team has been notified of this error, but you can also <0>tell us what went wrong</0>.","Tell us what went wrong":"Tell us what went wrong","between":"between","to":"to","when you\u2019re available for an appointment in the future":"when you\u2019re available for an appointment in the future"}};
