+++ web/locale/fr/messages.js
@@ -1,1 +1,1 @@
-module.exports={l:{"p":function(n,ord){if(ord)return n==1?"one":"other";return n>=0&&n<2?"one":"other"}},m:{"You can\u2019t select more than 3 days. To change your selections, remove some days first.":"Vous ne pouvez pas s\xE9lectionner plus de 3 dates. Pour modifier vos choix, veuillez d\u2019abord supprimer des dates.","Your 3 selected days:":"Vos 3 jours s\xE9lectionn\xE9s\xA0:","Remove day":"Supprimer cette journ\xE9e","Cancel request":"Annuler la demande","Please check these sections for errors:":"Veuillez v\xE9rifier les sections suivantes pour des erreurs\xA0:","Contact":"Contact","https://www.canada.ca/en/transparency/privacy.html":"https://www.canada.ca/fr/transparence/confidentialite.html","Privacy":"Confidentialit\xE9","https://digital.canada.ca/legal/terms/":"https://numerique.canada.ca/transparence/avis/","Terms":"Avis","Request a new citizenship appointment":"Demander un nouveau rendez-vous d\u2019examen de citoyennet\xE9","Language Selection":"Choix de la langue","This is a new service, help us improve by":"Il s\u2019agit d\u2019un nouveau service. Aidez-nous \xE0 l\u2019am\xE9liorer en nous","https://docs.google.com/forms/d/e/1FAIpQLSdEF3D7QCZ1ecPVKdqz_-dQAvlVdwdCQtHHLzg_v2q5q7XBlg/viewform":"https://docs.google.com/forms/d/e/1FAIpQLSesDeos-FqdEFaU8Bg0x9FXerEPausGX0svKKScYKw0l9US8g/viewform?hl=fr","sending your feedback":"envoyant vos commentaires","BETA":"B\xCATA","ALPHA":"ALPHA","Important message":"Message important","No days selected":"Aucune journ\xE9e s\xE9lectionn\xE9e","Change":"Modifier","Full name":"Nom complet","Email":"Courriel","Paper file number":"Num\xE9ro de dossier papier","Reason":"Raison","Explanation":"Explication","Availability":"Disponibilit\xE9","Travel":"Voyage","Calendar":"Calendrier","Go back":"Retour","Citizenship appointments are scheduled on Wednesdays and Thursdays.":"Les rendez-vous d\u2019examen de citoyennet\xE9 ont lieu les mercredis et les jeudis.","Select":"Choisissez","3 days":"3 jours","you\u2019re available between":"vous \xEAtes disponible entre","and":"et","Make sure you stay available on all of the days you select.":"Assurez-vous de demeurer disponible tous les jours s\xE9lectionn\xE9s.","Please select 2 more days to continue.":"Veuillez s\xE9lectionner 2 autres jours pour continuer.","Please select 1 more day to continue.":"Veuillez s\xE9lectionner 1 autre jour pour continuer.","Review request":"Revoir la demande","You must select 3 days.":"Vous devez s\xE9lectionner 3 jours.","Start over":"Recommencer","Your request has been cancelled.":"Votre demande a \xE9t\xE9 annul\xE9e.","Nothing about your application or appointment has been changed.":"Votre demande de citoyennet\xE9 ou votre rendez-vous n\u2019a pas \xE9t\xE9 modifi\xE9(e).","What happens next?":"Quelles sont les \xE9tapes suivantes?","Please attend your appointment at the date and time specified in the email we sent you.":"Veuillez vous pr\xE9senter \xE0 votre rendez-vous \xE0 la date et \xE0 l\u2019heure pr\xE9cis\xE9es dans le courriel que nous vous avons envoy\xE9.","If you missed your appointment, you still need to contact IRCC within 30 days. If you don\u2019t, your application could be delayed or abandoned.":"Si vous avez manqu\xE9 votre rendez-vous, vous devez tout de m\xEAme contacter IRCC d\u2019ici 30 jours, sinon le traitement de votre demande de citoyennet\xE9 pourrait \xEAtre retard\xE9 ou abandonn\xE9.","If you have any questions, please contact:":"Si vous avez des questions, communiquez avec\xA0:","Sorry, something went wrong. We received your request, but you might not get a confirmation email. Please make note of your request information":"D\xE9sol\xE9, une erreur s\u2019est produite. Nous avons re\xE7u votre demande, mais il se peut que vous ne recevrez pas le courriel de confirmation. Veuillez prendre note des renseignements de votre demande.","Availability:":"Disponibilit\xE9","Thank you! Your request has been received.":"Merci! Nous avons bien re\xE7u votre demande.","We\u2019ve sent you a confirmation email.":"Nous vous avons envoy\xE9 un courriel de confirmation.","By":"D\u2019ici le ","your local":"votre bureau local","office will send you a new appointment, or email you to ask for more information.":"vous enverra les d\xE9tails de votre nouveau rendez-vous ou communiquera avec vous par courriel pour vous demander des renseignements suppl\xE9mentaires.","Sorry, something went wrong.":"D\xE9sol\xE9, une erreur s\u2019est produite.","Your request":"Votre demande","was not completed":"n\u2019a pas \xE9t\xE9 compl\xE9t\xE9e","Your appointment or application wasn\u2019t changed in any way":"Votre demande de citoyennet\xE9 ou votre rendez-vous n\u2019a pas \xE9t\xE9 modifi\xE9(e).","Contact <0>IRCC</0> directly to reschedule your appointment:":"Veuillez communiquer directement avec IRCC pour reporter votre rendez-vous\xA0:","Our team has been notified of this error, but you can also <0>tell us what went wrong</0>.":"Notre \xE9quipe a \xE9t\xE9 avis\xE9e de l\u2019erreur, mais vous pouvez nous aider en d\xE9crivant ce qui s\u2019est pass\xE9.","Tell us what went wrong":"Veuillez d\xE9crire ce qui s\u2019est pass\xE9.","Home":"Accueil","Page not found":"Page non trouv\xE9e","Sorry, the page you are looking for doesn\u2019t exist.":"D\xE9sol\xE9, la page que vous cherchez n\u2019existe pas.","Return to the home page to reschedule your appointment.":"Veuillez revenir \xE0 la page d\u2019accueil pour reporter votre rendez-vous.","Tell IRCC you can\u2019t attend your citizenship appointment, and request a new one.":"Avisez IRCC que vous ne pouvez pas vous pr\xE9senter \xE0 votre rendez-vous d\u2019examen de citoyennet\xE9 et demandez un nouveau rendez-vous.","You will need:":"Vous devez\xA0:","Your paper file number":"Avoir en main votre num\xE9ro de dossier papier","To describe your reason for rescheduling":"Expliquer la raison du report","Then you\u2019ll select":"Ensuite, vous aurez \xE0 choisir","between":"entre","to":"\xE0","when you\u2019re available for an appointment in the future":"o\xF9 vous serez disponible pour vous pr\xE9senter \xE0 un rendez-vous","Requesting a new appointment will cancel your current one. Do not attend your old appointment after you complete this request. It can take up to 9 weeks for us to reschedule you.":"En demandant un nouveau rendez-vous, vous annulez votre rendez-vous actuel. Si vous soumettez cette demande, ne vous pr\xE9sentez pas \xE0 votre ancien rendez-vous. Le processus de report de votre rendez-vous peut nous prendre jusqu\u2019\xE0 9 semaines.","Start now":"Commencer","Email address":"Adresse courriel","Why are you rescheduling?":"Quelle est la raison du report?","Describe why you can\u2019t attend your appointment":"Veuillez d\xE9crire la raison pour laquelle vous ne pouvez pas vous pr\xE9senter \xE0 votre rendez-vous","Some information is missing.":"Il manque certains renseignements.","This is the full name you used on your citizenship application.":"Il s\u2019agit du nom complet que vous avez utilis\xE9 dans votre demande de citoyennet\xE9.","This is where we\u2019ll send a confirmation email when you\u2019re done.":"Il s\u2019agit de l\u2019adresse \xE0 laquelle nous enverrons un courriel de confirmation une fois que vous aurez compl\xE9t\xE9 le processus.","This number is at the top of the email attachment we sent you.":"Il s\u2019agit du num\xE9ro qui est affich\xE9 en haut du fichier que nous vous avons envoy\xE9 par courriel.","Travel (business or vacation)":"Voyage (d\u2019affaires ou vacances)","Medical":"Raisons m\xE9dicales","Work or school":"Travail ou \xE9tudes","Family":"Raisons familiales","Other":"Autre raison","Provide enough detail so that staff can understand your situation.":"Veuillez donner suffisament de d\xE9tails pour que le personnel puisse comprendre votre situation.","Continue":"Continuer","Send request":"Envoyer la demande","Work or School":"Travail ou \xE9tudes","Review your request:":"Revoir votre demande\xA0:","Sending this request will cancel your current appointment.<0> Do not attend your old appointment</0> after you send this request.":"Si vous envoyez cette demande, votre rendez-vous actuel sera annul\xE9. <0>Ne vous pr\xE9sentez pas \xE0 votre ancien rendez-vous</0> apr\xE8s avoir envoy\xE9 cette demande.","January":"janvier","February":"f\xE9vrier","March":"mars","April":"avril","May":"mai","June":"juin","July":"juillet","August":"ao\xFBt","September":"septembre","October":"octobre","November":"novembre","December":"d\xE9cembre","Sunday":"dimanche","Monday":"lundi","Tuesday":"mardi","Wednesday":"mercredi","Thursday":"jeudi","Friday":"vendredi","Saturday":"samedi","Su":"di","Mo":"lu","Tu":"ma","We":"me","Thu":"je","Fri":"ve","Sat":"sa","You need to tell us your name so we know who is requesting a new appointment.":"Nous avons besoin de conna\xEEtre votre nom pour savoir qui nous a soumis la demande de report.","Needs to be shorter than 20 words. Please use the name recorded on your application.":"Maximum de 20 mots. Veuillez utiliser le nom inclus dans votre demande de citoyennet\xE9.","We need your email address so we can send you a confirmation message.":"Nous avons besoin de votre adresse courriel afin de vous envoyer un message de confirmation.","Please make sure you provide a valid email address. For example, \u2018yourname@example.com\u2019.":"Veuillez vous assurez que l'adresse courriel est valide. Par exemple, \xAB\xA0votrenom@exemple.com\xA0\xBB.","We need your paper file number so we can confirm your identity.":"Nous avons besoin de votre num\xE9ro de dossier afin de confirmer votre identit\xE9.","Needs a number with a different format. Please make sure this is your correct Paper file number.":"Le num\xE9ro doit avoir un format diff\xE9rent. Veuillez v\xE9rifier que votre num\xE9ro de dossier papier est exact.","Please tell us why you need to reschedule your appointment. If none of the options fit your situation, choose \u2018Other\u2019.":"Veuillez nous fournir la raison pour laquelle vous voulez reporter votre rendez-vous. Si aucune des raisons \xE9num\xE9r\xE9es ne s\u2019applique \xE0 votre situation, choisissez \xAB\xA0Autre raison\xA0\xBB.","Please tell us a bit more about why you need to reschedule your appointment.":"Veuillez nous donner davantage de d\xE9tails sur votre raison de reporter votre rendez-vous.","Sorry, there\u2018s a limit of 150 words for this explanation. Please shorten your explanation.":"D\xE9sol\xE9, il y a une limite de 150 mots. Veuillez racourcir votre explication.","Exactly three dates must be passed":"Il faut s\xE9lectionner trois dates.","Reason needs to be on the list provided. Please pick one.":"La raison doit faire partie de la liste. Veuillez choisir une des options ci-dessous."}};
+module.exports={l:{"p":function(n,ord){if(ord)return n==1?"one":"other";return n>=0&&n<2?"one":"other"}},m:{"Immigration, Refugees and Citizenship Canada":"Immigration, R\xE9fugi\xE9s et Citoyennet\xE9 Canada","Our team has been notified of this error, but you can also":"Our team has been notified of this error, but you can also","let us know what happened":"let us know what happened","send an email to let us know what happened":"send an email to let us know what happened","You can\u2019t select more than 3 days. To change your selections, remove some days first.":"Vous ne pouvez pas s\xE9lectionner plus de 3 dates. Pour modifier vos choix, veuillez d\u2019abord supprimer des dates.","Your 3 selected days:":"Vos 3 jours s\xE9lectionn\xE9s\xA0:","Remove day":"Supprimer cette journ\xE9e","Cancel request":"Annuler la demande","Please check these sections for errors:":"Veuillez v\xE9rifier les sections suivantes pour des erreurs\xA0:","Contact":"Contact","https://www.canada.ca/en/transparency/privacy.html":"https://www.canada.ca/fr/transparence/confidentialite.html","Privacy":"Confidentialit\xE9","https://digital.canada.ca/legal/terms/":"https://numerique.canada.ca/transparence/avis/","Terms":"Avis","Request a new citizenship appointment":"Demander un nouveau rendez-vous d\u2019examen de citoyennet\xE9","Language Selection":"Choix de la langue","This is a new service, help us improve by":"Il s\u2019agit d\u2019un nouveau service. Aidez-nous \xE0 l\u2019am\xE9liorer en nous","https://docs.google.com/forms/d/e/1FAIpQLSdEF3D7QCZ1ecPVKdqz_-dQAvlVdwdCQtHHLzg_v2q5q7XBlg/viewform":"https://docs.google.com/forms/d/e/1FAIpQLSesDeos-FqdEFaU8Bg0x9FXerEPausGX0svKKScYKw0l9US8g/viewform?hl=fr","sending your feedback":"envoyant vos commentaires","BETA":"B\xCATA","ALPHA":"ALPHA","Important message":"Message important","No days selected":"Aucune journ\xE9e s\xE9lectionn\xE9e","Send request":"Envoyer la demande","Change":"Modifier","Full name":"Nom complet","Email":"Courriel","Paper file number":"Num\xE9ro de dossier papier","Reason":"Raison","Explanation":"Explication","Availability":"Disponibilit\xE9","Travel":"Voyage","Go back":"Retour","Citizenship appointments are scheduled on Wednesdays and Thursdays.":"Les rendez-vous d\u2019examen de citoyennet\xE9 ont lieu les mercredis et les jeudis.","Select":"Choisissez","3 days":"3 jours","you\u2019re available between":"vous \xEAtes disponible entre","and":"et","Make sure you stay available on all of the days you select.":"Assurez-vous de demeurer disponible tous les jours s\xE9lectionn\xE9s.","Please select 2 more days to continue.":"Veuillez s\xE9lectionner 2 autres jours pour continuer.","Please select 1 more day to continue.":"Veuillez s\xE9lectionner 1 autre jour pour continuer.","Review request":"Revoir la demande","You must select 3 days.":"Vous devez s\xE9lectionner 3 jours.","Start over":"Recommencer","Your request has been cancelled.":"Votre demande a \xE9t\xE9 annul\xE9e.","Nothing about your application or appointment has been changed.":"Votre demande de citoyennet\xE9 ou votre rendez-vous n\u2019a pas \xE9t\xE9 modifi\xE9(e).","What happens next?":"Quelles sont les \xE9tapes suivantes?","Please attend your appointment at the date and time specified in the email we sent you.":"Veuillez vous pr\xE9senter \xE0 votre rendez-vous \xE0 la date et \xE0 l\u2019heure pr\xE9cis\xE9es dans le courriel que nous vous avons envoy\xE9.","If you missed your appointment, you still need to contact IRCC within 30 days. If you don\u2019t, your application could be delayed or abandoned.":"Si vous avez manqu\xE9 votre rendez-vous, vous devez tout de m\xEAme contacter IRCC d\u2019ici 30 jours, sinon le traitement de votre demande de citoyennet\xE9 pourrait \xEAtre retard\xE9 ou abandonn\xE9.","If you have any questions, please contact:":"Si vous avez des questions, communiquez avec\xA0:","Sorry, something went wrong. We received your request, but you might not get a confirmation email. Please make note of your request information":"D\xE9sol\xE9, une erreur s\u2019est produite. Nous avons re\xE7u votre demande, mais il se peut que vous ne recevrez pas le courriel de confirmation. Veuillez prendre note des renseignements de votre demande.","Availability:":"Disponibilit\xE9","Thank you! Your request has been received.":"Merci! Nous avons bien re\xE7u votre demande.","We\u2019ve sent you a confirmation email.":"Nous vous avons envoy\xE9 un courriel de confirmation.","By":"D\u2019ici le ","your local":"votre bureau local","office will send you a new appointment, or email you to ask for more information.":"vous enverra les d\xE9tails de votre nouveau rendez-vous ou communiquera avec vous par courriel pour vous demander des renseignements suppl\xE9mentaires.","Sorry, something went wrong.":"D\xE9sol\xE9, une erreur s\u2019est produite.","Your request":"Votre demande","was not completed":"n\u2019a pas \xE9t\xE9 compl\xE9t\xE9e","Your appointment or application wasn\u2019t changed in any way":"Votre demande de citoyennet\xE9 ou votre rendez-vous n\u2019a pas \xE9t\xE9 modifi\xE9(e).","Contact <0>IRCC</0> directly to reschedule your appointment:":"Veuillez communiquer directement avec IRCC pour reporter votre rendez-vous\xA0:","Home":"Accueil","Page not found":"Page non trouv\xE9e","Sorry, the page you are looking for doesn\u2019t exist.":"D\xE9sol\xE9, la page que vous cherchez n\u2019existe pas.","Return to the home page to reschedule your appointment.":"Veuillez revenir \xE0 la page d\u2019accueil pour reporter votre rendez-vous.","Tell IRCC you can\u2019t attend your citizenship appointment, and request a new one.":"Avisez IRCC que vous ne pouvez pas vous pr\xE9senter \xE0 votre rendez-vous d\u2019examen de citoyennet\xE9 et demandez un nouveau rendez-vous.","You will need:":"Vous devez\xA0:","Your paper file number":"Avoir en main votre num\xE9ro de dossier papier","To describe your reason for rescheduling":"Expliquer la raison du report","Then you\u2019ll select":"Ensuite, vous aurez \xE0 choisir","Requesting a new appointment will cancel your current one. Do not attend your old appointment after you complete this request. It can take up to 9 weeks for us to reschedule you.":"En demandant un nouveau rendez-vous, vous annulez votre rendez-vous actuel. Si vous soumettez cette demande, ne vous pr\xE9sentez pas \xE0 votre ancien rendez-vous. Le processus de report de votre rendez-vous peut nous prendre jusqu\u2019\xE0 9 semaines.","Start now":"Commencer","Email address":"Adresse courriel","Why are you rescheduling?":"Quelle est la raison du report?","Describe why you can\u2019t attend your appointment":"Veuillez d\xE9crire la raison pour laquelle vous ne pouvez pas vous pr\xE9senter \xE0 votre rendez-vous","Some information is missing.":"Il manque certains renseignements.","This is the full name you used on your citizenship application.":"Il s\u2019agit du nom complet que vous avez utilis\xE9 dans votre demande de citoyennet\xE9.","This is where we\u2019ll send a confirmation email when you\u2019re done.":"Il s\u2019agit de l\u2019adresse \xE0 laquelle nous enverrons un courriel de confirmation une fois que vous aurez compl\xE9t\xE9 le processus.","This number is at the top of the email attachment we sent you.":"Il s\u2019agit du num\xE9ro qui est affich\xE9 en haut du fichier que nous vous avons envoy\xE9 par courriel.","Travel (business or vacation)":"Voyage (d\u2019affaires ou vacances)","Medical":"Raisons m\xE9dicales","Work or school":"Travail ou \xE9tudes","Family":"Raisons familiales","Other":"Autre raison","Provide enough detail so that staff can understand your situation.":"Veuillez donner suffisament de d\xE9tails pour que le personnel puisse comprendre votre situation.","Continue":"Continuer","Work or School":"Travail ou \xE9tudes","Review your request:":"Revoir votre demande\xA0:","Sending this request will cancel your current appointment.<0> Do not attend your old appointment</0> after you send this request.":"Si vous envoyez cette demande, votre rendez-vous actuel sera annul\xE9. <0>Ne vous pr\xE9sentez pas \xE0 votre ancien rendez-vous</0> apr\xE8s avoir envoy\xE9 cette demande.","January":"janvier","February":"f\xE9vrier","March":"mars","April":"avril","May":"mai","June":"juin","July":"juillet","August":"ao\xFBt","September":"septembre","October":"octobre","November":"novembre","December":"d\xE9cembre","Sunday":"dimanche","Monday":"lundi","Tuesday":"mardi","Wednesday":"mercredi","Thursday":"jeudi","Friday":"vendredi","Saturday":"samedi","Su":"di","Mo":"lu","Tu":"ma","We":"me","Thu":"je","Fri":"ve","Sat":"sa","You need to tell us your name so we know who is requesting a new appointment.":"Nous avons besoin de conna\xEEtre votre nom pour savoir qui nous a soumis la demande de report.","Needs to be shorter than 20 words. Please use the name recorded on your application.":"Maximum de 20 mots. Veuillez utiliser le nom inclus dans votre demande de citoyennet\xE9.","We need your email address so we can send you a confirmation message.":"Nous avons besoin de votre adresse courriel afin de vous envoyer un message de confirmation.","Please make sure you provide a valid email address. For example, \u2018yourname@example.com\u2019.":"Veuillez vous assurez que l'adresse courriel est valide. Par exemple, \xAB\xA0votrenom@exemple.com\xA0\xBB.","We need your paper file number so we can confirm your identity.":"Nous avons besoin de votre num\xE9ro de dossier afin de confirmer votre identit\xE9.","Needs a number with a different format. Please make sure this is your correct Paper file number.":"Le num\xE9ro doit avoir un format diff\xE9rent. Veuillez v\xE9rifier que votre num\xE9ro de dossier papier est exact.","Please tell us why you need to reschedule your appointment. If none of the options fit your situation, choose \u2018Other\u2019.":"Veuillez nous fournir la raison pour laquelle vous voulez reporter votre rendez-vous. Si aucune des raisons \xE9num\xE9r\xE9es ne s\u2019applique \xE0 votre situation, choisissez \xAB\xA0Autre raison\xA0\xBB.","Please tell us a bit more about why you need to reschedule your appointment.":"Veuillez nous donner davantage de d\xE9tails sur votre raison de reporter votre rendez-vous.","Sorry, there\u2018s a limit of 150 words for this explanation. Please shorten your explanation.":"D\xE9sol\xE9, il y a une limite de 150 mots. Veuillez racourcir votre explication.","Exactly three dates must be passed":"Il faut s\xE9lectionner trois dates.","Reason needs to be on the list provided. Please pick one.":"La raison doit faire partie de la liste. Veuillez choisir une des options ci-dessous.","Calendar":"Calendrier","Our team has been notified of this error, but you can also <0>tell us what went wrong</0>.":"Notre \xE9quipe a \xE9t\xE9 avis\xE9e de l\u2019erreur, mais vous pouvez nous aider en d\xE9crivant ce qui s\u2019est pass\xE9.","Tell us what went wrong":"Veuillez d\xE9crire ce qui s\u2019est pass\xE9.","between":"entre","to":"\xE0","when you\u2019re available for an appointment in the future":"o\xF9 vous serez disponible pour vous pr\xE9senter \xE0 un rendez-vous"}};
