+++ web/src/pages/ConfirmationPage.js
@@ -5,7 +5,7 @@ import styled, { css } from 'react-emotion'
 import { Trans } from 'lingui-react'
 import Layout from '../components/Layout'
 import Contact from '../components/Contact'
-import Abbreviation from '../components/abbreviation'
+import IRCCAbbr from '../components/IRCCAbbr'
 import { respondByDate } from '../utils/calendarDates'
 import withContext from '../withContext'
 import { contextPropTypes } from '../context'
@@ -104,7 +104,7 @@ class ConfirmationPage extends React.Component {
           </H2>
           <p>
             <Trans>By</Trans> {respondBy}
-            , <Trans>your local</Trans> <Abbreviation />{' '}
+            , <Trans>your local</Trans> <IRCCAbbr />{' '}
             <Trans>
               office will send you a new appointment, or email you to ask for
               more information.
